Move numpy int in dataId to explicit registry test
It was only incidentally being tested in the main
butler tests.

The numpy test is really a test to see that a numbers.Integral
object works with the dimensions system.  We don't really need
numpy for that but numpy is the easiest source of such a number.

Since this is a test and the only place in butler directly
using numpy, protect the import and skip the numbers.Integral
testing if numpy is missing.

It seems like far too much work to construct our own test
class that inherits from numbers.Integral that we can use
instead of numpy.

We explicitly do not test numpy.int64 values in dimension
definitions. These are not as user-facing as dataIds
and so it's much more likely that they will be correct
without requiring the overhead of having to test
the class of every single dimension.

+++ tests/test_butler.py
@@ -30,7 +30,6 @@ import shutil
 import pickle
 import string
 import random
-import numpy as np
 
 try:
     import boto3
@@ -868,7 +867,7 @@ class FileLikeDatastoreButlerTests(ButlerTests):
         butler.registry.registerDatasetType(DatasetType("metric2", dimensions, storageClass))
         butler.registry.registerDatasetType(DatasetType("metric3", dimensions, storageClass))
 
-        dataId1 = {"instrument": "DummyCamComp", "visit": np.int64(423)}
+        dataId1 = {"instrument": "DummyCamComp", "visit": 423}
         dataId2 = {"instrument": "DummyCamComp", "visit": 423, "physical_filter": "d-r"}
         dataId3 = {"instrument": "DummyCamComp", "visit": 425}
